Rename package Judge => JudgementController, JudgeMain.go => JudgementControllerMain.go
